requirements.txt
python-dotenv
aiogram==3.14
aiohttp==3.10.11